chores(diagram): 일부 함수 파일 이동, TODO 작성, lint 적용
